s2s.server/pyproject.toml
[project]
name = "s2s-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "torch",
    "transformers",
    "numpy",
    "scipy",
    "soundfile",
    "openai>=2.30.0",
    "google-genai>=1.73.1",
    "python-dotenv>=1.2.2",
]
